fix(auth): tighten snapshot integration
