+++ packages/auth/src/create-auth.ts
@@ -96,22 +96,6 @@ export type SessionStateAdapter = {
 	whenReady?: Promise<unknown>;
 };
 
-export type AuthWorkspaceSyncTarget = {
-	goOffline(): void;
-	reconnect(): void;
-};
-
-export type AuthWorkspaceTarget = {
-	sync: AuthWorkspaceSyncTarget;
-	idb: {
-		clearLocal(): Promise<unknown>;
-	};
-	encryption: {
-		applyKeys(keys: Session['encryptionKeys']): void;
-	};
-	getAuthSyncTargets?(): Iterable<AuthWorkspaceSyncTarget>;
-};
-
 export function createSessionStorageAdapter(
 	state: SessionStateAdapter,
 ): SessionStorage {
@@ -123,39 +107,6 @@ export function createSessionStorageAdapter(
 		save: (value) => state.set(value),
 		watch: state.watch,
 	};
-}
-
-export function attachAuthSnapshotToWorkspace({
-	auth,
-	workspace,
-	onSignedIn,
-}: {
-	auth: Pick<AuthClient, 'subscribe'>;
-	workspace: AuthWorkspaceTarget;
-	onSignedIn?: () => void;
-}): () => void {
-	function getSyncTargets() {
-		return new Set(workspace.getAuthSyncTargets?.() ?? [workspace.sync]);
-	}
-
-	return auth.subscribe((next, previous) => {
-		if (next.status === 'loading') return;
-
-		const previousSession =
-			previous.status === 'signedIn' ? previous.session : null;
-
-		if (next.status === 'signedOut') {
-			for (const sync of getSyncTargets()) sync.goOffline();
-			if (previousSession !== null) void workspace.idb.clearLocal();
-			return;
-		}
-
-		workspace.encryption.applyKeys(next.session.encryptionKeys);
-		if (previousSession?.token !== next.session.token) {
-			for (const sync of getSyncTargets()) sync.reconnect();
-		}
-		onSignedIn?.();
-	});
 }
 
 /**
@@ -268,13 +219,13 @@ export function createAuth({
 		});
 	}
 
-		function settleLoadedSession(loaded: Session | null) {
-			storageLoaded = true;
-			setSnapshot(snapshotFromSession(loaded));
-			if (bufferedBetterAuthCandidate !== undefined) {
-				reconcileBetterAuthCandidate(bufferedBetterAuthCandidate);
-			}
-			resolveWhenSessionLoaded();
+	function settleLoadedSession(loaded: Session | null) {
+		storageLoaded = true;
+		setSnapshot(snapshotFromSession(loaded));
+		if (bufferedBetterAuthCandidate !== undefined) {
+			reconcileBetterAuthCandidate(bufferedBetterAuthCandidate);
+		}
+		resolveWhenSessionLoaded();
 	}
 
 	function loadPersistedSession() {
