--- packages/svelte-utils/src/sync-status-popover/index.ts
@@ -1,1 +1,0 @@
-export { default as SyncStatusPopover } from './sync-status-popover.svelte';
